fix(ocpp): resolve connector_id type mismatch in StatusNotification upsert
The $2::text cast caused PostgreSQL to infer the parameter as text,
breaking the insert into the integer connector_id column.

+++ ocpp/v16/status/status.go
@@ -53,7 +53,7 @@ func upsertConnectorStatus(ctx context.Context, chargePointID string, connectorI
 		select
 			substr(md5(ec.id || '-' || $2::text), 1, 20),
 			ec.id,
-			$2,
+			$2::int,
 			$3,
 			$4,
 			$5,
